Update model tests  - Many models in test not converted to torch.bfloat16, fix it to convert  - Fix tests which run model is used as model.generate

+++ tests/models/albert/test_albert_masked_lm.py
@@ -8,10 +8,10 @@ from tests.utils import ModelTester
 
 class ThisTester(ModelTester):
     def _load_model(self):
-        return AlbertForMaskedLM.from_pretrained(self.model_name)
+        return AlbertForMaskedLM.from_pretrained(self.model_name, torch_dtype=torch.bfloat16)
 
     def _load_inputs(self):
-        self.tokenizer = AutoTokenizer.from_pretrained(self.model_name)
+        self.tokenizer = AutoTokenizer.from_pretrained(self.model_name, torch_dtype=torch.bfloat16)
         self.text = "The capital of [MASK] is Paris."
         self.inputs = self.tokenizer(self.text, return_tensors="pt")
         return self.inputs
